multiple settings modules added
