+++ demo/urls.py
@@ -2,11 +2,8 @@ from django.conf import settings
 from django.contrib import admin
 from django.urls import path, include
 
-from .views import home
-
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', home, name='home')
 ]
 
 if settings.DEBUG:
