--- demo/views.py
@@ -1,4 +1,0 @@
-from django.shortcuts import render
-
-def home(request):
-    return render(request, "index.html", {})
