--- templates/index.html
@@ -1,5 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <h1>THis is the home page</h1>
-{% endblock content %}
